Updated index.html and style.css
Added the solid background color along with the cards background color.
Added the image of the person for the card.
Added the font family.
Wrapped html content in tags.
Added CSS reset code.
Added classes to appropriate html.
Known issue: The background images for the page are not showing for unknown reasons.